Added a script and rule
Apple dictation is finiky. The biggest problem is the word int. The dictation does not understand this word and so fills it with things such as ain't or it's. I attempted researching a way to figure out how to add the word int to the dictation dictionary but I could only find how to add siri commands. I concluded I needed to keep moving forward and I added a rule stating to write the word int, the word integer must be stated. This functionality will be added next. I also created a script to know exactly what I should be stating for my example. After this though dictation is extremely unpredictable in how it interprets words, this may be an issue.