Updated the script and expected result to reflect swiftui language
Apple's dictation is a pain in my butt. It constantly gets words incorrect. I researched alternatives but I guess I should just move forward with it and note in the paper the issues it gives me. I can base my program off of the script and if apple doesn't transcribe it like the script or even similarly ( like 4 instead of for) then I will state so in the paper