Removed some specific functionality and continued creating functionality
I commented out the for loop functionality because I feel that may be too specific for the current time, but could be added back in later. I then updated the input example to reflect more written words and began updating the program to convert more written word to characters with special behaviors on quotes given it seems quotes are special. My next step is to add functionality for automatic tabbing.